Added CSS class to table, added/fixed comments
Added a new setting called "css_class_table", to allow a custom class to
be set on the table. If it's not defined, the table will be added a
hardcoded class named "ktable".

Renamed function "buildTable" to "buildTbody", as it is more correct.

Fixed minor layout bugs and typos in comments and added a couple of new
comments.